requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
google-generativeai
pydantic
pydantic-settings
python-dotenv
